[feat][refactor] More generic way of doing GET & POST requests that can be shared
